requirements.txt
SQLAlchemy==2.0.35
praw==7.8.1
prawcore==2.4.0
python-dotenv==1.0.1
email-validator==2.2.0

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
slowapi==0.1.9
python-multipart==0.0.17
jinja2==3.1.4

# Authentication
bcrypt==4.2.0
python-jose[cryptography]==3.3.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1
